Fix authentication in Edge (crypto API not supported inside web workers on Edge).
