+++ app/js/utils/workers/crypto-check.worker.js
@@ -1,0 +1,5 @@
+
+export async function isCryptoInWorkerSupported() {
+    const supported = global.crypto && global.crypto.getRandomValues;
+    return (!!supported).toString()
+}
